Refactor the packages and rewrite some methods
