+++ go/performance/import_benchmarker/helpers.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package import_benchmarker
 
 import (
 	"fmt"
